Repeats2.0_prerelease11

+++ Repeats/Pages/RepeatsList.xaml.cs
@@ -4,6 +4,7 @@ using Windows.UI.Xaml.Controls;
 using Microsoft.Data.Sqlite;
 using System.Threading.Tasks;
 using Windows.UI.Xaml.Media.Animation;
+using System.Linq;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
@@ -30,7 +31,9 @@ namespace Repeats.Pages
         {
             Button button = sender as Button;
             name = button.Tag.ToString();
-            Frame.Navigate(typeof(TakeTestPage));
+            Frame.Navigate(typeof(TakeTestPage), null, new DrillInNavigationTransitionInfo());
+
+            
         }
 
         public void ItemClick_Click(object sender, ItemClickEventArgs e)
@@ -60,53 +63,65 @@ namespace Repeats.Pages
             ContentDialogResult result = await ExcUPS.ShowAsync();
         }
 
-        private void Delete_Click(object sender, RoutedEventArgs e)
+        private async void Delete_Click(object sender, RoutedEventArgs e)
         {
             Button del = sender as Button;
             string gettag = del.Tag.ToString();
 
-            using (SqliteConnection db = new SqliteConnection("Filename=Repeats.db"))
+            ContentDialog delete = new ContentDialog
             {
-                db.Open();
-                String tableCommand = "DROP TABLE " + gettag;
-                SqliteCommand createTable = new SqliteCommand(tableCommand, db);
-                try
-                {
-                    createTable.ExecuteReader();
-                }
-                catch (SqliteException)
-                {
+                Title = "Czy na pewno chcesz usunąć ten zestaw?",
+                Content = "Usuwanie zestawów jest nieodwracalne.",
+                PrimaryButtonText = "Usuń",
+                CloseButtonText = "Anuluj"
+            };
 
+            ContentDialogResult result = await delete.ShowAsync();
+
+            if (result == ContentDialogResult.Primary)
+            {
+                ViewModel.Datas.Remove(new RepeatsListData() { TableName = gettag });
+
+                using (SqliteConnection db = new SqliteConnection("Filename=Repeats.db"))
+                {
+                    db.Open();
+                    String tableCommand = "DROP TABLE " + gettag;
+                    SqliteCommand createTable = new SqliteCommand(tableCommand, db);
+                    try
+                    {
+                        createTable.ExecuteReader();
+                    }
+                    catch (SqliteException)
+                    {
+
+                    }
+                }
+
+                using (SqliteConnection db = new SqliteConnection("Filename=Repeats.db"))
+                {
+                    db.Open();
+                    SqliteCommand insertCommand = new SqliteCommand();
+                    insertCommand.Connection = db;
+
+                    insertCommand.CommandText = "DELETE FROM TitleTable WHERE TableName=" + "\"" + gettag + "\"";
+                    try
+                    {
+                        insertCommand.ExecuteReader();
+                    }
+                    catch (SqliteException)
+                    {
+                        //Handle error
+                        return;
+                    }
+
+                    db.Close();
                 }
             }
-
-            using (SqliteConnection db = new SqliteConnection("Filename=Repeats.db"))
-            {
-                db.Open();
-                SqliteCommand insertCommand = new SqliteCommand();
-                insertCommand.Connection = db;
-
-                insertCommand.CommandText = "DELETE FROM TitleTable WHERE TableName=" + "\"" + gettag + "\"";
-                try
-                {
-                    insertCommand.ExecuteReader();
-                }
-                catch (SqliteException)
-                {
-                    //Handle error
-                    return;
-                }
-
-                db.Close();
-            }
-
-            ViewModel.Datas.Remove(new RepeatsListData() { TableName = gettag });
         }
-
 
         private void AddClick(object sender, RoutedEventArgs e)
         {
-            Frame.Navigate(typeof(AddRepeats));
+            Frame.Navigate(typeof(AddRepeats), null, new DrillInNavigationTransitionInfo());
         }
 
         private void BellClick(object sender, RoutedEventArgs e)
